pyproject.toml
[project]
name = "rockartdb"
version = "0.1.0"
description = "Rock art recording database with REST and GraphQL APIs"
authors = [
    {name = "James Williams",email = "james.williams@jtdub.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "graphene-django (>=2.16.0,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.5"
mkdocs = "^1.6.1"
black = "^25.11.0"
